move Theme and PreviewContainer to customview
