+++ app/build.gradle
@@ -50,6 +50,7 @@ android {
 
 dependencies {
     implementation(project(":data"))
+    implementation(project(":customview"))
 
     implementation platform(libs.compose.bom)
     implementation libs.compose.material3
