+++ app/src/main/java/com/github/kitakkun/noteapp/ui/page/editor/composable/EditorTopBar.kt
@@ -11,7 +11,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import com.github.kitakkun.noteapp.ui.preview.PreviewContainer
+import com.github.kitakkun.noteapp.customview.preview.PreviewContainer
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
