+++ app/src/main/java/com/github/kitakkun/noteapp/ui/page/editor/dialog/colorselect/SelectColorDialog.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
-import com.github.kitakkun.noteapp.ui.preview.PreviewContainer
+import com.github.kitakkun.noteapp.customview.preview.PreviewContainer
 
 @Composable
 fun SelectColorDialog(
