+++ app/src/main/java/com/github/kitakkun/noteapp/ui/page/finder/SearchTextField.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.github.kitakkun.noteapp.ui.preview.PreviewContainer
+import com.github.kitakkun.noteapp.customview.preview.PreviewContainer
 
 @Composable
 fun SearchField(
